macOS: move drag and drop handling into SurfaceView+DragDrop.swift
Pure code motion out of Ghostty.SurfaceView.swift; the only change is
dropping the MARK header the file name now carries.


# Conflicts:
#	macos/Sources/Ghostty/Surface View/Ghostty.SurfaceView.swift
